whatpm/t/xml/attrs-1.dat

#data
<a XYZ=""/>
#errors
#document
| <{} a>
|   XYZ=""

#data
<a lMn="" PQR=""/>
#errors
#document
| <{} a>
|   PQR=""
|   lMn=""

#data
<a b="" c=""/>
#errors
#document
| <{} a>
|   b=""
|   c=""

#data
<a><b c="" d=""/></a>
#errors
#document
| <{} a>
|   <{} b>
|     c=""
|     d=""

#data
<a/><b c="" d=""/>
#errors
1;5;second root element;;;m
#document
| <{} a>
| <{} b>
|   c=""
|   d=""
